CREST fleet call

+++ update/sql/2016-09-crest-fleet.sql
@@ -1,5 +1,5 @@
 create table crest_fleet (
-  id int(11) unsigned not null,
+  id bigint(11) unsigned not null,
   bossid int(11) unsigned not null,
   authgroupid int(11) unsigned not null,
   active tinyint(1) default 0,
